Many things done
Tutorial has a speaking npc, and walking to the left puts you in the main game.
all scenes fade in, I would like them to fade out but that's not so simple
I have confirmed that you can go through all scenes without it breaking.
the start menu is presentable.
the lights now flicker. we are free to adjust the effect as we see fit.